+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.0.3] - coming soon
+### Added
+* ArcGIS 10.5.1 / 2.0 projection codes
+  * 870 geographic coordinate system codes total (96 added)
+  * 6208 projected coordinate system codes (1456 added)
+  * 303 vertical codes (0 added)
+  * 2029 equation-based transformation codes (brand new!)
+  * 292 grid-based transformation codes (brand new!)
+  * 102 equation-based vertical transformation codes (brand new!)
+  * 53 grid-based transformation codes (brand new!)
+
 ## [1.0.2] - 2017-06-01
 ### Added
 * ArcGIS 10.5 projection codes
@@ -13,5 +24,6 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 * Support for browserify and other bundlers (see Esri/esri-proj-codes#4 for more info)
 
+[1.0.3]: https://github.com/Esri/esri-proj-codes/compare/v1.0.2...v1.0.3
 [1.0.2]: https://github.com/Esri/esri-proj-codes/compare/v1.0.1...v1.0.2
 [1.0.1]: https://github.com/Esri/esri-proj-codes/compare/v1.0.0...v1.0.1